backend/requirements.txt
fastapi
uvicorn
gunicorn
requests
cryptography
redis
fastapi-limiter
pydantic
pydantic-settings
python-jose
passlib
python-multipart
aiohttp
mangum
pytest
pytest-asyncio
httpx
psycopg2-binary>=2.9.3
sqlalchemy>=1.4.0
pyotp
qrcode[pil]
pillow
python-dotenv
argon2-cffi
sib_api_v3_sdk